fix: Removed black lines between tiles while moving
The reason had been, that everything is drawn relativ to the camera
position. But the camera position had been a float value. So while
moving it occoured that the computer could not render the correct
position of a tile. By truncating the camera position, the values
are always easy to draw for the computer. And the flickers are gone
Cheers ;)